+++ src/app/(route)/(monster)/monster/[slug]/result/page.tsx
@@ -8,7 +8,7 @@ import { LayoutGroup } from 'framer-motion';
 
 import CharacterCanvas from '@components/CharacterCanvas';
 import Button from '@components/common/Button';
-import Toast from '@components/common/feedback/Toast';
+import Toast from '@components/common/Toast';
 
 import useModal from '@hooks/useModal';
