+++ src/app/(route)/(monster)/monster/[slug]/share/components/ExpiredModal.tsx
@@ -1,6 +1,6 @@
 import { useRouter } from 'next/navigation';
 
-import Modal from '@components/common/feedback/Modal';
+import Modal from '@components/common/Modal';
 
 type ExpiredModalProps = {
   closeModal?: () => void;
