+++ src/app/components/common/Button/index.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { Slot } from '@radix-ui/react-slot';
 import { cva, VariantProps } from 'class-variance-authority';
 
-import Icon from '@components/common/data-display/Icon';
+import Icon from '@components/common/Icon';
 
 import cn from '@utils/cn';
