+++ src/app/components/common/navigation/Menu/MenuItem.tsx
@@ -2,7 +2,7 @@ import Link from 'next/link';
 
 import type { ComponentPropsWithoutRef, ElementType } from 'react';
 
-import Icon from '@components/common/data-display/Icon';
+import Icon from '@components/common/Icon';
 
 import cn from '@utils/cn';
