+++ src/app/components/modals/SaveMonsterImageModal.tsx
@@ -6,7 +6,7 @@ import html2canvas from 'html2canvas';
 import CharacterCanvas, {
   type CharacterCanvasProps,
 } from '@components/CharacterCanvas';
-import Modal from '@components/common/feedback/Modal';
+import Modal from '@components/common/Modal';
 
 type SaveMonsterImageModalProps = {
   monster: CharacterCanvasProps;
